Python interface started

+++ src/fortplot.f90
@@ -3,20 +3,20 @@ module fortplot
     !!
     !! This module provides a clean, user-friendly API for creating scientific plots
     !! with support for line plots, contour plots, and multiple output formats.
-    !! 
+    !!
     !! Quick Start:
     !!   use fortplot
     !!   type(figure_t) :: fig
-    !!   
+    !!
     !!   call fig%initialize(640, 480)
     !!   call fig%add_plot(x, y, label="data")
     !!   call fig%savefig('output.png')
     !!
     !! Author: fortplotlib contributors
 
+    use iso_fortran_env, only: wp => real64
     use fortplot_figure, only: figure_t
     use fortplot_format_parser, only: parse_format_string, contains_format_chars
-    use, intrinsic :: iso_fortran_env, only: wp => real64
 
     implicit none
 
@@ -29,7 +29,7 @@ module fortplot
     public :: savefig, figure
     public :: add_plot, add_contour, add_contour_filled
     public :: set_xscale, set_yscale
-    
+
     ! Line style constants (pyplot-style)
     character(len=*), parameter, public :: LINESTYLE_SOLID = '-'
     character(len=*), parameter, public :: LINESTYLE_DASHED = '--'
@@ -40,15 +40,15 @@ module fortplot
     ! Marker style constants (pyplot-style)
     character(len=*), parameter, public :: MARKER_CIRCLE = 'o'
     character(len=*), parameter, public :: MARKER_CROSS = 'x'
-    
-    
+
+
     ! Interface for overloaded show routine
     interface show
         module procedure show_data, show_figure
     end interface show
-    
+
     ! Note: plot and add_plot now handle both separate args and format strings in single routines
-    
+
     ! Global figure for simple API
     type(figure_t), save :: fig
 
@@ -62,9 +62,9 @@ contains
         !!   label: Optional label for the plot
         !!   linestyle: Line style and markers ('b-o', 'r--', 'g:', 'ko', etc.)
         !!              Supports both pyplot-fortran and matplotlib format strings
-        real(wp), dimension(:), intent(in) :: x, y
+        real(8), dimension(:), intent(in) :: x, y
         character(len=*), intent(in), optional :: label, linestyle
-        
+
         call fig%add_plot(x, y, label=label, linestyle=linestyle)
     end subroutine plot
 
@@ -76,11 +76,11 @@ contains
         !!   z: 2D data array for contouring
         !!   levels: Optional array of contour levels
         !!   label: Optional label for the plot
-        real(wp), dimension(:), intent(in) :: x, y
-        real(wp), dimension(:,:), intent(in) :: z
-        real(wp), dimension(:), intent(in), optional :: levels
+        real(8), dimension(:), intent(in) :: x, y
+        real(8), dimension(:,:), intent(in) :: z
+        real(8), dimension(:), intent(in), optional :: levels
         character(len=*), intent(in), optional :: label
-        
+
         call fig%add_contour(x, y, z, levels=levels, label=label)
     end subroutine contour
 
@@ -94,12 +94,12 @@ contains
         !!   colormap: Optional colormap name ('crest' (default), 'viridis', 'plasma', 'rocket', 'mako', 'flare', etc.)
         !!   show_colorbar: Optional flag to show colorbar
         !!   label: Optional label for the plot
-        real(wp), dimension(:), intent(in) :: x, y
-        real(wp), dimension(:,:), intent(in) :: z
-        real(wp), dimension(:), intent(in), optional :: levels
+        real(8), dimension(:), intent(in) :: x, y
+        real(8), dimension(:,:), intent(in) :: z
+        real(8), dimension(:), intent(in), optional :: levels
         character(len=*), intent(in), optional :: colormap, label
         logical, intent(in), optional :: show_colorbar
-        
+
         call fig%add_contour_filled(x, y, z, levels=levels, colormap=colormap, &
                                    show_colorbar=show_colorbar, label=label)
     end subroutine contour_filled
@@ -113,15 +113,15 @@ contains
         !!   label: Optional label for the plot
         !!   title_text: Optional plot title
         !!   xlabel_text, ylabel_text: Optional axis labels
-        real(wp), dimension(:), intent(in) :: x, y
+        real(8), dimension(:), intent(in) :: x, y
         character(len=*), intent(in), optional :: label, title_text, xlabel_text, ylabel_text
-        
+
         call fig%initialize()
-        
+
         if (present(title_text)) call fig%set_title(title_text)
         if (present(xlabel_text)) call fig%set_xlabel(xlabel_text)
         if (present(ylabel_text)) call fig%set_ylabel(ylabel_text)
-        
+
         call fig%add_plot(x, y, label=label)
         call fig%show()
     end subroutine show_data
@@ -134,11 +134,11 @@ contains
 
     subroutine figure(width, height)
         !! Initialize the global figure for simple API usage
-        !!  
+        !!
         !! Arguments:
         !!   width, height: Optional figure dimensions (default: 640x480)
         integer, intent(in), optional :: width, height
-        
+
         if (present(width) .and. present(height)) then
             call fig%initialize(width, height)
         else
@@ -179,26 +179,26 @@ contains
 
     subroutine add_plot(x, y, label, linestyle)
         !! Add a line plot to the global figure (pyplot-fortran compatible)
-        real(wp), dimension(:), intent(in) :: x, y
+        real(8), dimension(:), intent(in) :: x, y
         character(len=*), intent(in), optional :: label, linestyle
-        
+
         call fig%add_plot(x, y, label=label, linestyle=linestyle)
     end subroutine add_plot
 
     subroutine add_contour(x, y, z, levels, label)
         !! Add a contour plot to the global figure
-        real(wp), dimension(:), intent(in) :: x, y
-        real(wp), dimension(:,:), intent(in) :: z
-        real(wp), dimension(:), intent(in), optional :: levels
+        real(8), dimension(:), intent(in) :: x, y
+        real(8), dimension(:,:), intent(in) :: z
+        real(8), dimension(:), intent(in), optional :: levels
         character(len=*), intent(in), optional :: label
         call fig%add_contour(x, y, z, levels=levels, label=label)
     end subroutine add_contour
 
     subroutine add_contour_filled(x, y, z, levels, colormap, show_colorbar, label)
         !! Add a filled contour plot with color levels to the global figure
-        real(wp), dimension(:), intent(in) :: x, y
-        real(wp), dimension(:,:), intent(in) :: z
-        real(wp), dimension(:), intent(in), optional :: levels
+        real(8), dimension(:), intent(in) :: x, y
+        real(8), dimension(:,:), intent(in) :: z
+        real(8), dimension(:), intent(in), optional :: levels
         character(len=*), intent(in), optional :: colormap, label
         logical, intent(in), optional :: show_colorbar
         call fig%add_contour_filled(x, y, z, levels=levels, colormap=colormap, &
@@ -208,14 +208,14 @@ contains
     subroutine set_xscale(scale, threshold)
         !! Set x-axis scale for the global figure
         character(len=*), intent(in) :: scale
-        real(wp), intent(in), optional :: threshold
+        real(8), intent(in), optional :: threshold
         call fig%set_xscale(scale, threshold)
     end subroutine set_xscale
 
     subroutine set_yscale(scale, threshold)
-        !! Set y-axis scale for the global figure  
+        !! Set y-axis scale for the global figure
         character(len=*), intent(in) :: scale
-        real(wp), intent(in), optional :: threshold
+        real(8), intent(in), optional :: threshold
         call fig%set_yscale(scale, threshold)
     end subroutine set_yscale
